requirements.txt
streamlit>=1.30.0
pandas>=2.0.0
plotly>=5.18.0
python-docx>=1.1.0
pyotp>=2.9.0
qrcode>=7.4.2
supabase>=2.3.0
